+++ android/src/main/java/com/mahatech/flutter_pax_printer_utility/PrinterUtility.java
@@ -250,34 +250,34 @@ public class PrinterUtility {
         String res = "";
         switch (status) {
             case 0:
-                res = "Success ";
+                res = "Success";
                 break;
             case 1:
-                res = "Printer is busy ";
+                res = "Printer is busy";
                 break;
             case 2:
-                res = "Out of paper ";
+                res = "Out of paper";
                 break;
             case 3:
-                res = "The format of print data packet error ";
+                res = "The format of print data packet error";
                 break;
             case 4:
-                res = "Printer malfunctions ";
+                res = "Printer malfunctions";
                 break;
             case 8:
-                res = "Printer over heats ";
+                res = "Printer over heats";
                 break;
             case 9:
                 res = "Printer voltage is too low";
                 break;
             case 240:
-                res = "Printing is unfinished ";
+                res = "Printing is unfinished";
                 break;
             case 252:
-                res = " The printer has not installed font library ";
+                res = "The printer has not installed font library";
                 break;
             case 254:
-                res = "Data package is too long ";
+                res = "Data package is too long";
                 break;
             default:
                 break;