+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,8 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import flutter_pax_printer_utility
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  FlutterPaxPrinterUtilityPlugin.register(with: registry.registrar(forPlugin: "FlutterPaxPrinterUtilityPlugin"))
 }